@@ -18,7 +18,8 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.8.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter_sdk: ">=2.0.0"
 
 dependencies:
   flutter:
@@ -27,43 +28,43 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
+  cupertino_icons: ^1.0.3
   
   # realiza la comparacion de objetos automaticamente (equals())
-  equatable: ^1.2.0
+  equatable: ^2.0.3
   # permite realizar llamadas http
-  http: ^0.12.1
+  http: ^0.13.3
   # permite realizar un return entre dos tipos de clases y actuar a corde
-  dartz: ^0.9.2
+  dartz: ^0.10.0
   # para di(dependency injection)
-  get_it: ^5.0.1
+  get_it: ^7.2.0
   # google fonts
-  google_fonts: ^1.1.1
+  google_fonts: ^2.1.0
   # bloc design pattern
-  flutter_bloc: ^6.1.1
+  flutter_bloc: ^7.3.0
   # handle svgs files
-  flutter_svg: ^0.18.1
+  flutter_svg: ^0.22.0
   # to load images from internet
-  cached_network_image: ^2.2.0+1
+  cached_network_image: ^3.1.0
   # to i18n
   flutter_localizations:
     sdk: flutter
   # feedback module
-  wiredash: ^0.2.0
+  wiredash: ^0.7.0+1
   # playing/stream YouTube videos inline
-  youtube_player_flutter: ^7.0.0+7
+  youtube_player_flutter: ^8.0.0
   # local dba
-  hive: ^1.4.1+1
-  path_provider: ^1.6.11
+  hive: ^2.0.4
+  path_provider: ^2.0.5
   # animations on rive app
-  flare_flutter: ^2.0.6
+  flare_flutter: ^3.0.2
 
 dev_dependencies:
-  hive_generator: 0.7.2
-  build_runner: ^1.10.0
-  test: ^1.15.4
-  bloc_test: ^7.1.0
-  mockito: ^4.1.2
+  hive_generator: ^1.1.1
+  build_runner: ^2.1.2
+  test: ^1.17.10
+  bloc_test: ^8.2.0
+  mockito: ^5.0.16
   flutter_test:
     sdk: flutter
 
