@@ -1,39 +1,39 @@
-
 import 'package:movie_app/domain/entities/movie_entity.dart';
 
 class MovieModel extends MovieEntity {
-  final String releaseDate;
-  final bool adult;
-  final String backdropPath;
-  final List<int> genreIds;
-  final int voteCount;
-  final String originalLanguage;
-  final String originalTitle;
-  final String posterPath;
-  final String title;
-  final bool video;
-  final double voteAverage;
   final int id;
-  final String overview;
-  final double popularity;
-  final String mediaType;
+  final bool? video;
+  final int? voteCount;
+  final double? voteAverage;
+  final String title;
+  final String? releaseDate;
+  final String? originalLanguage;
+  final String? originalTitle;
+  final List<int>? genreIds;
+  final String backdropPath;
+  final bool? adult;
+  final String? overview;
+  final String posterPath;
+  final double? popularity;
+  final String? mediaType;
 
-  MovieModel(
-      {this.releaseDate,
-      this.adult,
-      this.backdropPath,
-      this.genreIds,
-      this.voteCount,
-      this.originalLanguage,
-      this.originalTitle,
-      this.posterPath,
-      this.title,
-      this.video,
-      this.voteAverage,
-      this.id,
-      this.overview,
-      this.popularity,
-      this.mediaType}) : super(
+  MovieModel({
+    required this.id,
+    this.video,
+    this.voteCount,
+    this.voteAverage,
+    required this.title,
+    this.releaseDate,
+    this.originalLanguage,
+    this.originalTitle,
+    this.genreIds,
+    required this.backdropPath,
+    this.adult,
+    this.overview,
+    required this.posterPath,
+    this.popularity,
+    this.mediaType,
+  }) : super(
           id: id,
           title: title,
           backdropPath: backdropPath,
@@ -45,39 +45,39 @@ class MovieModel extends MovieEntity {
 
   factory MovieModel.fromJson(Map<String, dynamic> json) {
     return MovieModel(
-      releaseDate: json['release_date'],
-      adult: json['adult'],
-      backdropPath: json['backdrop_path'],
-      genreIds: json['genre_ids'].cast<int>(),
+      popularity: json['popularity']?.toDouble() ?? 0.0,
       voteCount: json['vote_count'],
+      video: json['video'],
+      posterPath: json['poster_path'] ?? '',
+      id: json['id'] ?? -1,
+      adult: json['adult'],
+      backdropPath: json['backdrop_path'] ?? '',
       originalLanguage: json['original_language'],
       originalTitle: json['original_title'],
-      posterPath: json['poster_path'],
-      title: json['title'],
-      video: json['video'],
+      genreIds: json['genre_ids'].cast<int>(),
+      title: json['title'] ?? '',
       voteAverage: json['vote_average']?.toDouble() ?? 0.0,
-      id: json['id'],
       overview: json['overview'],
-      popularity: json['popularity']?.toDouble() ?? 0.0,
-      mediaType: json['media_type']
+      releaseDate: json['release_date'],
+      mediaType: json['media_type'],
     );
   }
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
-    data['release_date'] = this.releaseDate;
-    data['adult'] = this.adult;
-    data['backdrop_path'] = this.backdropPath;
-    data['genre_ids'] = this.genreIds;
+    data['id'] = this.id;
+    data['video'] = this.video;
     data['vote_count'] = this.voteCount;
+    data['vote_average'] = this.voteAverage;
+    data['title'] = this.title;
+    data['release_date'] = this.releaseDate;
     data['original_language'] = this.originalLanguage;
     data['original_title'] = this.originalTitle;
-    data['poster_path'] = this.posterPath;
-    data['title'] = this.title;
-    data['video'] = this.video;
-    data['vote_average'] = this.voteAverage;
-    data['id'] = this.id;
+    data['genre_ids'] = this.genreIds;
+    data['backdrop_path'] = this.backdropPath;
+    data['adult'] = this.adult;
     data['overview'] = this.overview;
+    data['poster_path'] = this.posterPath;
     data['popularity'] = this.popularity;
     data['media_type'] = this.mediaType;
     return data;
