@@ -1,27 +1,35 @@
 import 'video_model.dart';
 
 class VideoResultModel {
-  int id;
-  List<VideoModel> videos;
+  final int id;
+  late final List<VideoModel> videos;
 
-  VideoResultModel({this.id, this.videos});
+  VideoResultModel({required this.id, required this.videos});
 
-  VideoResultModel.fromJson(Map<String, dynamic> json) {
-    id = json['id'];
+  factory VideoResultModel.fromJson(Map<String, dynamic> json) {
+    var videos = List<VideoModel>.empty(growable: true);
     if (json['results'] != null) {
-      videos = new List<VideoModel>();
       json['results'].forEach((v) {
-        videos.add(new VideoModel.fromJson(v));
+        var _videoModel = VideoModel.fromJson(v);
+        if (_isValidVideo(_videoModel)) {
+          videos.add(VideoModel.fromJson(v));
+        }
       });
     }
+
+    return VideoResultModel(id: json['id'], videos: videos);
   }
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
     data['id'] = this.id;
-    if (this.videos != null) {
-      data['results'] = this.videos.map((v) => v.toJson()).toList();
-    }
+    data['results'] = this.videos.map((v) => v.toJson()).toList();
     return data;
   }
+}
+
+bool _isValidVideo(VideoModel videoModel) {
+  return videoModel.key.isNotEmpty &&
+      videoModel.name.isNotEmpty &&
+      videoModel.type.isNotEmpty;
 }
