@@ -1,8 +1,9 @@
-import 'package:flutter/material.dart';
-
 class LanguageEntity {
   final String code;
   final String value;
 
-  const LanguageEntity({@required this.code, @required this.value});
+  const LanguageEntity({
+    required this.code,
+    required this.value,
+  });
 }
