@@ -1,5 +1,4 @@
 import 'package:equatable/equatable.dart';
-import 'package:flutter/foundation.dart';
 import 'package:movie_app/domain/entities/movie_detail_entity.dart';
 
 class MovieEntity extends Equatable {
@@ -7,19 +6,19 @@ class MovieEntity extends Equatable {
   final int id;
   final String backdropPath;
   final String title;
-  final num voteAverage;
-  final String releaseDate;
-  final String overview;
+  final num? voteAverage;
+  final String? releaseDate;
+  final String? overview;
 
   const MovieEntity({
-    @required this.posterPath,
-    @required this.id,
-    @required this.backdropPath,
-    @required this.title,
-    @required this.voteAverage,
-    @required this.releaseDate,
+    required this.posterPath,
+    required this.id,
+    required this.backdropPath,
+    required this.title,
+    required this.voteAverage,
+    required this.releaseDate,
     this.overview,
-  }) : assert(id != null, 'Movie id must not be null');
+  });
 
   @override
   List<Object> get props => [id, title];
@@ -32,7 +31,7 @@ class MovieEntity extends Equatable {
     return MovieEntity(
       posterPath: movieDetailEntity.posterPath,
       id: movieDetailEntity.id,
-      backdropPath: movieDetailEntity.backdropPath,
+      backdropPath: movieDetailEntity.backdropPath ?? '',
       title: movieDetailEntity.title,
       voteAverage: movieDetailEntity.voteAverage,
       releaseDate: movieDetailEntity.releaseDate,
