@@ -1,7 +1,6 @@
 import 'package:bloc/bloc.dart';
 import 'package:dartz/dartz.dart';
 import 'package:equatable/equatable.dart';
-import 'package:flutter/material.dart';
 import 'package:movie_app/domain/entities/app_error.dart';
 import 'package:movie_app/domain/entities/cast_entity.dart';
 import 'package:movie_app/domain/entities/movie_params.dart';
@@ -12,7 +11,9 @@ part 'cast_state.dart';
 class CastCubit extends Cubit<CastState> {
   final GetCast getCast;
 
-  CastCubit({@required this.getCast}) : super(CastInitial());
+  CastCubit({
+    required this.getCast,
+  }) : super(CastInitial());
 
   void loadCast(int movieId) async {
     Either<AppError, List<CastEntity>> eitherResponse =
