@@ -1,5 +1,4 @@
 import 'package:dartz/dartz.dart';
-import 'package:flutter/material.dart';
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
 import 'package:movie_app/common/constants/translation_constants.dart';
@@ -18,9 +17,9 @@ class LoginCubit extends Cubit<LoginState> {
   final LoadingCubit loadingCubit;
 
   LoginCubit({
-    @required this.loginUser,
-    @required this.logoutUser,
-    @required this.loadingCubit,
+    required this.loginUser,
+    required this.logoutUser,
+    required this.loadingCubit,
   }) : super(LoginInitial());
 
   void login(String username, String password) async {
