@@ -1,6 +1,5 @@
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
-import 'package:flutter/material.dart';
 import 'package:movie_app/domain/entities/app_error.dart';
 
 import 'package:movie_app/domain/entities/movie_entity.dart';
@@ -17,9 +16,9 @@ class MovieCarouselCubit extends Cubit<MovieCarouselState> {
   final LoadingCubit loadingCubit;
 
   MovieCarouselCubit({
-    @required this.getTrending,
-    @required this.movieBackdropCubit,
-    @required this.loadingCubit,
+    required this.getTrending,
+    required this.movieBackdropCubit,
+    required this.loadingCubit,
   }) : super(MovieCarouselInitial());
 
   void loadCarousel({int movieIndex = 0}) async {
