@@ -1,7 +1,6 @@
 import 'package:bloc/bloc.dart';
 import 'package:dartz/dartz.dart';
 import 'package:equatable/equatable.dart';
-import 'package:flutter/material.dart';
 
 import 'package:movie_app/domain/entities/app_error.dart';
 import 'package:movie_app/domain/entities/movie_detail_entity.dart';
@@ -22,11 +21,11 @@ class MovieDetailCubit extends Cubit<MovieDetailState> {
   final LoadingCubit loadingCubit;
 
   MovieDetailCubit({
-    @required this.loadingCubit,
-    @required this.castCubit,
-    @required this.getMovieDetail,
-    @required this.videosCubit,
-    @required this.favouriteCubit,
+    required this.loadingCubit,
+    required this.castCubit,
+    required this.getMovieDetail,
+    required this.videosCubit,
+    required this.favouriteCubit,
   }) : super(MovieDetailInitial());
 
   void loadMovieDetail(int movieId) async {
