@@ -1,7 +1,6 @@
 import 'package:bloc/bloc.dart';
 import 'package:dartz/dartz.dart';
 import 'package:equatable/equatable.dart';
-import 'package:flutter/material.dart';
 
 import 'package:movie_app/domain/entities/app_error.dart';
 import 'package:movie_app/domain/entities/movie_params.dart';
@@ -13,7 +12,9 @@ part 'videos_state.dart';
 class VideosCubit extends Cubit<VideosState> {
   final GetVideos getVideos;
 
-  VideosCubit({@required this.getVideos}) : super(VideosInitial());
+  VideosCubit({
+    required this.getVideos,
+  }) : super(VideosInitial());
 
   void loadVideos(int movieId) async {
     final Either<AppError, List<VideoEntity>> eitherVideoResponse =
