@@ -14,7 +14,9 @@ class NoVideos extends VideosState {}
 class VideosLoaded extends VideosState {
   final List<VideoEntity> videos;
 
-  VideosLoaded({@required this.videos});
+  VideosLoaded({
+    required this.videos,
+  });
 
   @override
   List<Object> get props => [videos];
