@@ -1,21 +1,20 @@
 import 'package:flutter/material.dart';
+import 'package:movie_app/common/constants/size_constants.dart';
 
 import 'package:movie_app/common/extensions/size_extensions.dart';
 
 class NavigationListItem extends StatelessWidget {
-
   final String title;
-  final Function onPressed;
+  final Function() onPressed;
 
   const NavigationListItem({
-    Key key,
-    @required this.title,
-    @required this.onPressed
+    Key? key,
+    required this.title,
+    required this.onPressed,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-
     return GestureDetector(
       onTap: onPressed,
       child: Container(
@@ -40,10 +39,13 @@ class NavigationListItem extends StatelessWidget {
 
 class NavigationSubListItem extends StatelessWidget {
   final String title;
-  final Function onPressed;
+  final Function() onPressed;
 
-  const NavigationSubListItem({Key key, this.title, this.onPressed})
-      : super(key: key);
+  const NavigationSubListItem({
+    Key? key,
+    required this.title,
+    required this.onPressed,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -59,7 +61,7 @@ class NavigationSubListItem extends StatelessWidget {
           ],
         ),
         child: ListTile(
-          contentPadding: EdgeInsets.symmetric(horizontal: 32.w),
+          contentPadding: EdgeInsets.symmetric(horizontal: Sizes.dimen_32.w),
           title: Text(
             title,
             style: Theme.of(context).textTheme.subtitle1,
