@@ -12,17 +12,19 @@ import 'package:movie_app/presentation/journeys/home/movie_carousel/movie_carous
 import 'package:movie_app/presentation/journeys/home/movie_tabbed/movie_tabbed_widget.dart';
 
 class HomeScreen extends StatefulWidget {
-  HomeScreen({Key key}) : super(key: key);
+  HomeScreen({
+    Key? key,
+  }) : super(key: key);
 
   @override
   _HomeScreenState createState() => _HomeScreenState();
 }
 
 class _HomeScreenState extends State<HomeScreen> {
-  MovieCarouselCubit movieCarouselCubit;
-  MovieBackdropCubit movieBackdropCubit;
-  MovieTabbedCubit movieTabbedCubit;
-  SearchMovieCubit searchMovieCubit;
+  late MovieCarouselCubit movieCarouselCubit;
+  late MovieBackdropCubit movieBackdropCubit;
+  late MovieTabbedCubit movieTabbedCubit;
+  late SearchMovieCubit searchMovieCubit;
 
   @override
   void initState() {
@@ -37,10 +39,10 @@ class _HomeScreenState extends State<HomeScreen> {
   @override
   void dispose() {
     super.dispose();
-    movieCarouselCubit?.close();
-    movieBackdropCubit?.close();
-    movieTabbedCubit?.close();
-    searchMovieCubit?.close();
+    movieCarouselCubit.close();
+    movieBackdropCubit.close();
+    movieTabbedCubit.close();
+    searchMovieCubit.close();
   }
 
   @override
@@ -55,7 +57,7 @@ class _HomeScreenState extends State<HomeScreen> {
       child: Scaffold(
         drawer: const NavigationDrawer(),
         body: BlocBuilder<MovieCarouselCubit, MovieCarouselState>(
-          cubit: movieCarouselCubit,
+          bloc: movieCarouselCubit,
           builder: (context, state) {
             if (state is MovieCarouselLoaded) {
               return Stack(
