@@ -7,35 +7,35 @@ import 'package:movie_app/common/screenutil/screenutil.dart';
 import 'movie_card_widget.dart';
 
 class AnimatedMovieCardWidget extends StatelessWidget {
-
   final int index;
   final PageController pageController;
   final int movieId;
   final String posterPath;
 
   const AnimatedMovieCardWidget({
-    Key key,
-    @required this.index,
-    @required this.movieId,
-    @required this.posterPath,
-    @required this.pageController,
+    Key? key,
+    required this.index,
+    required this.movieId,
+    required this.posterPath,
+    required this.pageController,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    
     return AnimatedBuilder(
       animation: pageController,
       builder: (context, child) {
         double value = 1;
 
         if (pageController.position.haveDimensions) {
-          value = pageController.page - index;
+          value = (pageController.page ?? 0) - index;
           value = (1 - (value.abs() * 0.1)).clamp(0.0, 1.0);
           return Align(
             alignment: Alignment.topCenter,
             child: Container(
-              height: Curves.easeIn.transform(value) * ScreenUtil.screenHeight * 0.35,
+              height: Curves.easeIn.transform(value) *
+                  ScreenUtil.screenHeight *
+                  0.35,
               width: Sizes.dimen_230.w,
               child: child,
             ),
@@ -46,7 +46,8 @@ class AnimatedMovieCardWidget extends StatelessWidget {
             child: Container(
               height:
                   Curves.easeIn.transform(index == 0 ? value : value * 0.5) *
-                      ScreenUtil.screenHeight * 0.35,
+                      ScreenUtil.screenHeight *
+                      0.35,
               width: Sizes.dimen_230.w,
               child: child,
             ),
