@@ -12,9 +12,9 @@ class MovieCardWidget extends StatelessWidget {
   final String posterPath;
 
   const MovieCardWidget({
-    Key key,
-    @required this.movieId,
-    @required this.posterPath,
+    Key? key,
+    required this.movieId,
+    required this.posterPath,
   }) : super(key: key);
 
   @override
