@@ -8,16 +8,15 @@ import 'movie_data_widget.dart';
 import 'movie_page_view.dart';
 
 class MovieCarouselWidget extends StatelessWidget {
-
   final List<MovieEntity> movies;
   final int defaultIndex;
-  
+
   const MovieCarouselWidget({
-      Key key,
-      @required this.movies,
-      @required this.defaultIndex,
-    })  : assert(defaultIndex >= 0, 'defaultIndex cannot be less than 0'),
-          super(key: key);
+    Key? key,
+    required this.movies,
+    required this.defaultIndex,
+  })  : assert(defaultIndex >= 0, 'defaultIndex cannot be less than 0'),
+        super(key: key);
 
   @override
   Widget build(BuildContext context) {
