@@ -4,8 +4,8 @@ import 'package:flutter/material.dart';
 class LoadingCircle extends StatelessWidget {
   final double size;
   const LoadingCircle({
-    Key key,
-    @required this.size,
+    Key? key,
+    required this.size,
   }) : super(key: key);
 
   @override
