@@ -6,7 +6,9 @@ import 'package:movie_app/presentation/widgets/logo.dart';
 import 'login_form.dart';
 
 class LoginScreen extends StatelessWidget {
-  const LoginScreen({Key key}) : super(key: key);
+  const LoginScreen({
+    Key? key,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
