@@ -14,7 +14,10 @@ import 'movie_detail_app_bar.dart';
 class BigPoster extends StatelessWidget {
   final MovieDetailEntity movie;
 
-  const BigPoster({Key key, @required this.movie}) : super(key: key);
+  const BigPoster({
+    Key? key,
+    required this.movie,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -43,7 +46,7 @@ class BigPoster extends StatelessWidget {
             style: Theme.of(context).textTheme.headline5,
           ),
           subtitle: Text(
-            movie.releaseDate,
+            movie.releaseDate ?? '',
             style: Theme.of(context).textTheme.greySubtitle1,
           ),
           trailing: Text(
