@@ -3,15 +3,13 @@ import 'package:flutter/material.dart';
 import 'package:movie_app/common/extensions/size_extensions.dart';
 
 class Logo extends StatelessWidget {
-
   final double height;
 
   const Logo({
-    Key key,
-    @required this.height
-  }) : assert(height != null, 'heihgt must not be null'),
-      assert(height > 0, 'height should be greater than 0'),
-      super(key: key);
+    Key? key,
+    required this.height,
+  })  : assert(height > 0, 'height should be greater than 0'),
+        super(key: key);
 
   @override
   Widget build(BuildContext context) {
