@@ -1,4 +1,4 @@
-extension NumExtension on num {
+extension NumExtension on num? {
   String convertToPercentageString() {
     return ((this ?? 0) * 10).toStringAsFixed(0) + ' %';
   }
