@@ -1,9 +1,9 @@
 import 'package:movie_app/common/screenutil/screenutil.dart';
 
-extension SizeExtension on num {
-  num get w => ScreenUtil().setWidth(this);
+extension SizeExtension on double {
+  double get w => ScreenUtil().setWidth(this);
 
-  num get h => ScreenUtil().setHeight(this);
+  double get h => ScreenUtil().setHeight(this);
 
-  num get sp => ScreenUtil().setSp(this);
+  double get sp => ScreenUtil().setSp(this);
 }
