@@ -5,8 +5,8 @@ extension StringExtension on String {
   String intelliTrim() {
     return this.length > 15 ? '${this.substring(0, 15)}...' : this;
   }
-  
- String translate(BuildContext context) {
-  return AppLocalizations.of(context).translate(this);
- }
+
+  String translate(BuildContext context) {
+    return AppLocalizations.of(context)?.translate(this) ?? '';
+  }
 }
