@@ -9,7 +9,7 @@ class ApiClient {
 
   ApiClient(this._client);
 
-  dynamic get(String path, {Map<dynamic, dynamic> params}) async {
+  dynamic get(String path, {Map<dynamic, dynamic>? params}) async {
     final response = await _client.get(
       getPath(path, params),
       headers: {
@@ -24,7 +24,7 @@ class ApiClient {
     }
   }
 
-  dynamic post(String path, {Map<dynamic, dynamic> params}) async {
+  dynamic post(String path, {Map<dynamic, dynamic>? params}) async {
     final response = await _client.post(
       getPath(path, null),
       body: jsonEncode(params),
@@ -42,8 +42,11 @@ class ApiClient {
     }
   }
 
-  dynamic deleteWithBody(String path, {Map<dynamic, dynamic> params}) async {
-    Request request = Request('DELETE', Uri.parse(getPath(path, null)));
+  dynamic deleteWithBody(String path, {Map<dynamic, dynamic>? params}) async {
+    Request request = Request(
+      'DELETE',
+      getPath(path, null),
+    );
     request.headers['Content-Type'] = 'application/json';
     request.body = jsonEncode(params);
     final response = await _client.send(request).then(
@@ -59,14 +62,15 @@ class ApiClient {
     }
   }
 
-  String getPath(String path, Map<dynamic, dynamic> params) {
+  Uri getPath(String path, Map<dynamic, dynamic>? params) {
     var paramsString = '';
     if (params?.isNotEmpty ?? false) {
-      params.forEach((key, value) {
+      params?.forEach((key, value) {
         paramsString += '&$key=$value';
       });
     }
 
-    return '${ApiConstants.BASE_URL}$path?api_key=${ApiConstants.API_KEY}&$paramsString';
+    return Uri.parse(
+        '${ApiConstants.BASE_URL}$path?api_key=${ApiConstants.API_KEY}&$paramsString');
   }
 }
