@@ -4,7 +4,7 @@ import 'package:movie_app/data/models/request_token_model.dart';
 abstract class AuthenticationRemoteDataSource {
   Future<RequestTokenModel> getRequestToken();
   Future<RequestTokenModel> validateWithLogin(Map<String, dynamic> requestBody);
-  Future<String> createSession(Map<String, dynamic> requestBody);
+  Future<String?> createSession(Map<String, dynamic> requestBody);
   Future<bool> deleteSession(String sessionId);
 }
 
@@ -15,7 +15,7 @@ class AuthenticationRemoteDataSourceImpl
   AuthenticationRemoteDataSourceImpl(this._client);
 
   @override
-  Future<String> createSession(Map<String, dynamic> requestBody) async {
+  Future<String?> createSession(Map<String, dynamic> requestBody) async {
     final response = await _client.post(
       'authentication/session/new',
       params: requestBody,
