@@ -2,18 +2,22 @@ import 'package:equatable/equatable.dart';
 
 class MovieDetailEntity extends Equatable {
   final int id;
-  final String title, releaseDate, overview, posterPath, backdropPath;
-  final num voteAverage;
+  final String title;
+  final String? releaseDate;
+  final String? overview;
+  final String posterPath;
+  final String? backdropPath;
+  final num? voteAverage;
 
-  const MovieDetailEntity(
-    this.id,
-    this.title,
-    this.releaseDate,
-    this.overview,
-    this.voteAverage,
-    this.backdropPath,
-    this.posterPath
-  );
+  const MovieDetailEntity({
+    required this.id,
+    required this.title,
+    required this.releaseDate,
+    required this.overview,
+    required this.voteAverage,
+    required this.backdropPath,
+    required this.posterPath,
+  });
 
   @override
   List<Object> get props => [id];
