@@ -2,7 +2,9 @@ import 'package:flutter/cupertino.dart';
 import 'package:movie_app/common/constants/routes_constants.dart';
 import 'package:movie_app/presentation/journeys/favourite/favourite_screen.dart';
 import 'package:movie_app/presentation/journeys/home/home_screen.dart';
+import 'package:movie_app/presentation/journeys/movie_detail/movie_detail_arguments.dart';
 import 'package:movie_app/presentation/journeys/movie_detail/movie_detail_screen.dart';
+import 'package:movie_app/presentation/journeys/watch_video/watch_video_arguments.dart';
 import 'package:movie_app/presentation/journeys/watch_video/watch_video_screen.dart';
 
 import 'journeys/login/login_screen.dart';
@@ -12,10 +14,10 @@ class Routes {
         RouteList.initial: (context) => LoginScreen(),
         RouteList.home: (context) => HomeScreen(),
         RouteList.movieDetail: (context) => MovieDetailScreen(
-              movieDetailArguments: settings.arguments,
+              movieDetailArguments: settings.arguments as MovieDetailArguments,
             ),
         RouteList.watchTrailer: (context) => WatchVideoScreen(
-              watchVideoArguments: settings.arguments,
+              watchVideoArguments: settings.arguments as WatchVideoArguments,
             ),
         RouteList.favourite: (context) => FavouriteScreen(),
       };
